requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
backtesting==0.6.4
bokeh==3.7.3
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
dateparser==1.2.1
dotenv==0.9.9
frozenlist==1.6.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.3
narwhals==1.39.1
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
pycryptodome==3.23.0
python-binance==1.0.28
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
tornado==6.5
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
websockets==15.0.1
xyzservices==2025.4.0
yarl==1.20.0
